+++ fbw-a32nx/src/systems/fmgc/src/flightplanning/plans/BaseFlightPlan.ts
@@ -853,11 +853,11 @@ export abstract class BaseFlightPlan<P extends FlightPlanPerformanceData = Fligh
     }
   }
 
-  get originAirport(): Airport {
+  get originAirport(): Airport | undefined {
     return this.originSegment.originAirport;
   }
 
-  async setOriginAirport(icao: string) {
+  async setOriginAirport(icao: string | undefined) {
     await this.originSegment.setAirport(icao);
 
     await this.flushOperationQueue();
@@ -868,7 +868,7 @@ export abstract class BaseFlightPlan<P extends FlightPlanPerformanceData = Fligh
     return this.originSegment.originRunway;
   }
 
-  async setOriginRunway(runwayIdent: string) {
+  async setOriginRunway(runwayIdent: string | undefined) {
     await this.originSegment.setRunway(runwayIdent);
 
     await this.flushOperationQueue();
@@ -969,7 +969,7 @@ export abstract class BaseFlightPlan<P extends FlightPlanPerformanceData = Fligh
     this.incrementVersion();
   }
 
-  get destinationAirport(): Airport {
+  get destinationAirport(): Airport | undefined {
     return this.destinationSegment.destinationAirport;
   }