+++ fbw-a32nx/src/systems/fmgc/src/flightplanning/plans/FlightPlan.ts
@@ -1,5 +1,4 @@
-// @ts-strict-ignore
-// Copyright (c) 2021-2025 FlyByWire Simulations
+// Copyright (c) 2021-2026 FlyByWire Simulations
 // Copyright (c) 2021-2022 Synaptic Simulations
 //
 // SPDX-License-Identifier: GPL-3.0
@@ -8,7 +7,6 @@ import { Airport, ApproachType, Fix, isMsfs2024, LegType, MathUtils, NXDataStore
 import { AlternateFlightPlan } from '@fmgc/flightplanning/plans/AlternateFlightPlan';
 import { AeroMath, BitFlags, EventBus, MagVar, MutableSubscribable } from '@microsoft/msfs-sdk';
 import { FixInfoData, FixInfoEntry } from '@fmgc/flightplanning/plans/FixInfo';
-import { loadAllDepartures, loadAllRunways } from '@fmgc/flightplanning/DataLoading';
 import { Coordinates, Degrees } from 'msfs-geo';
 import { FlightPlanLeg, FlightPlanLegFlags } from '@fmgc/flightplanning/legs/FlightPlanLeg';
 import { SegmentClass } from '@fmgc/flightplanning/segments/SegmentClass';
@@ -49,7 +47,7 @@ export class FlightPlan<P extends FlightPlanPerformanceData = FlightPlanPerforma
   /**
    * FIX INFO entries
    */
-  fixInfos: readonly FixInfoEntry[] = [];
+  fixInfos: readonly (FixInfoEntry | undefined)[] = [];
 
   /**
    * Shown as the "flight number" in the MCDU, but it's really the callsign
@@ -110,40 +108,47 @@ export class FlightPlan<P extends FlightPlanPerformanceData = FlightPlanPerforma
     return newPlan;
   }
 
-  get alternateDestinationAirport(): Airport {
+  get alternateDestinationAirport(): Airport | undefined {
     return this.alternateFlightPlan.destinationAirport;
   }
 
   async setAlternateDestinationAirport(icao: string | undefined) {
     await this.deleteAlternateFlightPlan();
+    await this.alternateFlightPlan.setOriginAirport(this.destinationAirport?.ident);
     await this.alternateFlightPlan.setDestinationAirport(icao);
-
-    if (this.alternateFlightPlan.originAirport) {
-      this.alternateFlightPlan.availableOriginRunways = await loadAllRunways(this.alternateFlightPlan.originAirport);
-      this.alternateFlightPlan.availableDepartures = await loadAllDepartures(this.alternateFlightPlan.originAirport);
-    }
-
-    await this.alternateFlightPlan.originSegment.refreshOriginLegs();
 
     await this.alternateFlightPlan.flushOperationQueue();
   }
 
   async deleteAlternateFlightPlan() {
-    await this.alternateFlightPlan.setOriginRunway(undefined);
-    await this.alternateFlightPlan.setDeparture(undefined);
+    // unset procedures in an order least likely to leave legs behind
     await this.alternateFlightPlan.setDepartureEnrouteTransition(undefined);
-    await this.alternateFlightPlan.setDestinationRunway(undefined);
+    await this.alternateFlightPlan.setDeparture(undefined);
+    await this.alternateFlightPlan.setOriginRunway(undefined);
     await this.alternateFlightPlan.setArrivalEnrouteTransition(undefined);
     await this.alternateFlightPlan.setArrival(undefined);
-    await this.alternateFlightPlan.setApproach(undefined);
     await this.alternateFlightPlan.setApproachVia(undefined);
+    await this.alternateFlightPlan.setApproach(undefined);
+    await this.alternateFlightPlan.setDestinationRunway(undefined);
+
+    // delete this last in case any other operations put legs in enroute
+    this.alternateFlightPlan.enrouteSegment.clear();
+
+    // and finally get rid of the airports (stuff might try to use them while legs still exist)
     await this.alternateFlightPlan.setDestinationAirport(undefined);
+    await this.alternateFlightPlan.setOriginAirport(undefined);
 
     this.alternateFlightPlan.allLegs.length = 0;
 
     this.resetAlternatePerformanceData();
 
     this.alternateFlightPlan.incrementVersion();
+
+    if (this.alternateFlightPlan.allLegs.length > 0) {
+      console.warn('[FlightPlan::deleteAlternateFlightPlan] Legs left over after clearing the plan!', [
+        ...this.alternateFlightPlan.allLegs,
+      ]);
+    }
   }
 
   private resetAlternatePerformanceData() {
@@ -162,7 +167,8 @@ export class FlightPlan<P extends FlightPlanPerformanceData = FlightPlanPerforma
     }
 
     const targetLeg = this.legElementAt(targetLegIndex);
-    if (!targetLeg.isXF()) {
+    const targetLegFix = targetLeg.terminationWaypoint();
+    if (!targetLeg.isXF() || !targetLegFix) {
       throw new Error('[FPM] Cannot direct to a non-XF leg');
     }
 
@@ -175,7 +181,7 @@ export class FlightPlan<P extends FlightPlanPerformanceData = FlightPlanPerforma
       turningPoint.flags |= FlightPlanLegFlags.PendingDirectToTurningPoint;
     }
 
-    const turnEnd = FlightPlanLeg.directToTurnEnd(this.enrouteSegment, targetLeg.terminationWaypoint())
+    const turnEnd = FlightPlanLeg.directToTurnEnd(this.enrouteSegment, targetLegFix)
       .withDefinitionFrom(targetLeg)
       .withPilotEnteredDataFrom(targetLeg);
     // If we don't do this, the turn end will have the termination waypoint's ident which may not be the leg ident (for runway legs for example)
@@ -268,7 +274,11 @@ export class FlightPlan<P extends FlightPlanPerformanceData = FlightPlanPerforma
       const [segment, xfLegIndexInSegment] = this.segmentPositionForIndex(xFLegIndexInPlan);
       const xfLegAfterDiscontinuity = segment.allLegs[xfLegIndexInSegment] as FlightPlanLeg;
 
-      if (xfLegAfterDiscontinuity.type !== LegType.IF && xfLegAfterDiscontinuity.type !== LegType.CF) {
+      if (
+        xfLegAfterDiscontinuity.type !== LegType.IF &&
+        xfLegAfterDiscontinuity.type !== LegType.CF &&
+        xfLegAfterDiscontinuity.definition.waypoint
+      ) {
         const iFLegAfterDiscontinuity = FlightPlanLeg.fromEnrouteFix(
           segment,
           xfLegAfterDiscontinuity.definition.waypoint,
@@ -343,7 +353,6 @@ export class FlightPlan<P extends FlightPlanPerformanceData = FlightPlanPerforma
     );
 
     this.deleteAlternateFlightPlan();
-    this.resetAlternatePerformanceData();
 
     this.enqueueOperation(FlightPlanQueuedOperation.RebuildArrivalAndApproach);
     this.enqueueOperation(FlightPlanQueuedOperation.Restring);
@@ -357,7 +366,7 @@ export class FlightPlan<P extends FlightPlanPerformanceData = FlightPlanPerforma
   }
 
   setFixInfoEntry(index: 1 | 2 | 3 | 4, fixInfo: FixInfoData | null, notify = true): void {
-    const planFixInfo = this.fixInfos as FixInfoEntry[];
+    const planFixInfo = this.fixInfos as (FixInfoEntry | undefined)[];
 
     planFixInfo[index] = fixInfo ? new FixInfoEntry(fixInfo.fix, fixInfo?.radii, fixInfo?.radials) : undefined;
 
@@ -448,7 +457,11 @@ export class FlightPlan<P extends FlightPlanPerformanceData = FlightPlanPerforma
   async setDestinationAirport(icao: string | undefined): Promise<void> {
     await super.setDestinationAirport(icao);
 
-    FlightPlan.setDestinationDefaultPerformanceData(this, this.destinationAirport);
+    if (this.destinationAirport) {
+      FlightPlan.setDestinationDefaultPerformanceData(this, this.destinationAirport);
+    }
+
+    return this.alternateFlightPlan.setOriginAirport(icao);
   }
 
   /**
@@ -632,22 +645,21 @@ export class FlightPlan<P extends FlightPlanPerformanceData = FlightPlanPerforma
    */
   reconcileThrustReductionWithConstraints(): boolean {
     const lowestClimbConstraint = MathUtils.round(this.lowestClimbConstraint(), 10);
-    if (
-      Number.isFinite(lowestClimbConstraint) &&
-      this.performanceData.thrustReductionAltitude.get() !== null &&
-      this.performanceData.thrustReductionAltitude.get() > lowestClimbConstraint
-    ) {
+    const thrRed = this.performanceData.thrustReductionAltitude.get();
+
+    if (Number.isFinite(lowestClimbConstraint) && thrRed !== null && thrRed > lowestClimbConstraint) {
+      const defaultThrRed = this.performanceData.defaultThrustReductionAltitude.get();
+
       this.setPerformanceData(
         'defaultThrustReductionAltitude',
-        this.performanceData.defaultThrustReductionAltitude.get() !== null
-          ? Math.min(this.performanceData.defaultThrustReductionAltitude.get(), lowestClimbConstraint)
-          : null,
+        defaultThrRed !== null ? Math.min(defaultThrRed, lowestClimbConstraint) : null,
       );
+
+      const pilotThrRed = this.performanceData.pilotThrustReductionAltitude.get();
+
       this.setPerformanceData(
         'pilotThrustReductionAltitude',
-        this.performanceData.pilotThrustReductionAltitude.get() !== null
-          ? Math.min(this.performanceData.pilotThrustReductionAltitude.get(), lowestClimbConstraint)
-          : null,
+        pilotThrRed !== null ? Math.min(pilotThrRed, lowestClimbConstraint) : null,
       );
 
       return true;
@@ -662,22 +674,20 @@ export class FlightPlan<P extends FlightPlanPerformanceData = FlightPlanPerforma
    */
   reconcileAccelerationWithConstraints(): boolean {
     const lowestClimbConstraint = MathUtils.round(this.lowestClimbConstraint(), 10);
-    if (
-      Number.isFinite(lowestClimbConstraint) &&
-      this.performanceData.accelerationAltitude.get() !== null &&
-      this.performanceData.accelerationAltitude.get() > lowestClimbConstraint
-    ) {
+    const accAlt = this.performanceData.accelerationAltitude.get();
+
+    if (Number.isFinite(lowestClimbConstraint) && accAlt !== null && accAlt > lowestClimbConstraint) {
+      const defaultAccAlt = this.performanceData.defaultAccelerationAltitude.get();
+
       this.setPerformanceData(
         'defaultAccelerationAltitude',
-        this.performanceData.defaultAccelerationAltitude.get() !== null
-          ? Math.min(this.performanceData.defaultAccelerationAltitude.get(), lowestClimbConstraint)
-          : null,
+        defaultAccAlt !== null ? Math.min(defaultAccAlt, lowestClimbConstraint) : null,
       );
+
+      const pilotAccAlt = this.performanceData.pilotAccelerationAltitude.get();
       this.setPerformanceData(
         'pilotAccelerationAltitude',
-        this.performanceData.pilotAccelerationAltitude.get() !== null
-          ? Math.min(this.performanceData.pilotAccelerationAltitude.get(), lowestClimbConstraint)
-          : null,
+        pilotAccAlt !== null ? Math.min(pilotAccAlt, lowestClimbConstraint) : null,
       );
 
       return true;